Renamed imports to AuthHelpers
